added subscription types to worker, dependent destroy to trackable models

+++ app/controllers/subscriptions_controller.rb
@@ -12,7 +12,7 @@ class SubscriptionsController < ApplicationController
   end
 
   def destroy
-    subscription = Subscription.where(subscribable_id: params[:id]).delete_all
+    subscription = Subscription.where(user_id: current_user.id, subscribable_id: params[:id]).delete_all
     redirect_to :back, notice: t('messages.subscriptions.destroy')
   end
